AV1/requirements.txt
numpy
pandas
scikit-learn
spacy
kagglehub
nltk
pyspellchecker
gensim
huggingface_hub
safetensors